I don't get paid enough for this shit.
Fix some stuff. Maybe.

+++ include/tun.h
@@ -31,7 +31,7 @@ void		 tnt_ttc_close(struct device *);
 int		 tnt_ttc_set_ip(struct device *, const char *);
 int		 tnt_ttc_up(struct device *);
 int		 tnt_ttc_down(struct device *);
-int		 tnt_ttc_get_fd(struct device *);
+intptr_t		 tnt_ttc_get_fd(struct device *);
 
 #endif
